backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.2
requests==2.31.0
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.10
langchain-pinecone==0.0.1
pypdf==3.17.4
pinecone-client==3.0.0
supabase==2.3.0
